+++ packages/provider-utils/src/types/infer-tool-context.ts
@@ -1,7 +1,12 @@
+import { HasRequiredKey } from '../has-required-key';
 import type { Tool } from './tool';
 
 /**
  * Infer the context type of a tool.
  */
-export type InferToolContext<TOOL extends Tool<any, any, any>> =
-  TOOL extends Tool<any, any, infer CONTEXT> ? CONTEXT : never;
+export type InferToolContext<TOOL extends Tool> =
+  TOOL extends Tool<any, any, infer CONTEXT>
+    ? HasRequiredKey<CONTEXT> extends true
+      ? CONTEXT
+      : never
+    : never;