+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/azure
 
+## 4.0.0-beta.28
+
+### Patch Changes
+
+- Updated dependencies [3ae1786]
+  - @ai-sdk/provider-utils@5.0.0-beta.17
+  - @ai-sdk/openai@4.0.0-beta.28
+
 ## 4.0.0-beta.27
 
 ### Patch Changes