+++ packages/google/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/google
 
+## 4.0.0-beta.33
+
+### Patch Changes
+
+- Updated dependencies [3ae1786]
+  - @ai-sdk/provider-utils@5.0.0-beta.17
+
 ## 4.0.0-beta.32
 
 ### Patch Changes