+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ### Patch Changes
 
+- Updated dependencies [2fe1099]
+- Updated dependencies [f04adcb]
+  - ai@7.0.0-beta.81
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3ae1786]
+  - ai@7.0.0-beta.80
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [6866afe]
   - ai@7.0.0-beta.79